refactor(foundation): move PromotionStatus/Request/History types (A12)
Phase 6 / Sub-phase A12: Extract more workspace_promotion types — now that
MockEnvironmentName, PromotionEntityType, PromotionService are in foundation,
the data types that compose them can move too.

Foundation:
- workspace_promotion::PromotionStatus (Pending, Approved, Rejected, Completed, Failed)
- workspace_promotion::PromotionRequest (full struct matching core's API)
- workspace_promotion::PromotionHistory
- workspace_promotion::PromotionHistoryEntry

Core re-export shims in workspace/scenario_promotion.rs.

mockforge-collab:
- src/promotion.rs migrated to import MockEnvironmentName + PromotionService
  from foundation
- Added mockforge-foundation Cargo dep

Workspace builds clean.
