+++ crates/mockforge-core/src/workspace/scenario_promotion.rs
@@ -11,31 +11,8 @@ use std::collections::HashMap;
 // `PromotionEntityType` is re-exported from `mockforge_foundation::workspace_promotion`.
 pub use mockforge_foundation::workspace_promotion::PromotionEntityType;
 
-/// Generic promotion request that supports scenarios, personas, and configs
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct PromotionRequest {
-    /// Entity type being promoted
-    pub entity_type: PromotionEntityType,
-    /// Entity ID to promote (scenario ID, persona ID, or "config" for config promotion)
-    pub entity_id: String,
-    /// Entity version (for scenarios/personas) or config snapshot ID (for configs)
-    pub entity_version: Option<String>,
-    /// Workspace ID
-    pub workspace_id: String,
-    /// Source environment
-    pub from_environment: MockEnvironmentName,
-    /// Target environment
-    pub to_environment: MockEnvironmentName,
-    /// Whether this requires approval
-    pub requires_approval: bool,
-    /// Reason why approval is required
-    pub approval_required_reason: Option<String>,
-    /// Comments from promoter
-    pub comments: Option<String>,
-    /// Additional metadata for the promotion (e.g., config changes diff)
-    #[serde(default)]
-    pub metadata: HashMap<String, serde_json::Value>,
-}
+// `PromotionRequest` is re-exported from `mockforge_foundation::workspace_promotion`.
+pub use mockforge_foundation::workspace_promotion::PromotionRequest;
 
 /// Scenario promotion request (backward compatibility)
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -90,33 +67,8 @@ pub struct ScenarioPromotionResult {
     pub status: PromotionStatus,
 }
 
-/// Promotion status
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "lowercase")]
-pub enum PromotionStatus {
-    /// Promotion is pending approval
-    Pending,
-    /// Promotion has been approved
-    Approved,
-    /// Promotion has been rejected
-    Rejected,
-    /// Promotion has been completed
-    Completed,
-    /// Promotion failed
-    Failed,
-}
-
-impl std::fmt::Display for PromotionStatus {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            PromotionStatus::Pending => write!(f, "pending"),
-            PromotionStatus::Approved => write!(f, "approved"),
-            PromotionStatus::Rejected => write!(f, "rejected"),
-            PromotionStatus::Completed => write!(f, "completed"),
-            PromotionStatus::Failed => write!(f, "failed"),
-        }
-    }
-}
+// `PromotionStatus` is re-exported from `mockforge_foundation::workspace_promotion`.
+pub use mockforge_foundation::workspace_promotion::PromotionStatus;
 
 /// Scenario promotion workflow manager
 ///
@@ -386,50 +338,11 @@ impl CustomApprovalRule {
     }
 }
 
-/// Promotion history for an entity (scenario, persona, or config)
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct PromotionHistory {
-    /// Entity type
-    pub entity_type: PromotionEntityType,
-    /// Entity ID
-    pub entity_id: String,
-    /// Workspace ID
-    pub workspace_id: String,
-    /// List of promotions in chronological order
-    pub promotions: Vec<PromotionHistoryEntry>,
-}
+// `PromotionHistory` is re-exported from `mockforge_foundation::workspace_promotion`.
+pub use mockforge_foundation::workspace_promotion::PromotionHistory;
 
-/// Single promotion history entry
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct PromotionHistoryEntry {
-    /// Promotion ID
-    pub promotion_id: String,
-    /// Entity type
-    pub entity_type: PromotionEntityType,
-    /// Entity ID
-    pub entity_id: String,
-    /// Entity version (for scenarios/personas) or config snapshot ID (for configs)
-    pub entity_version: Option<String>,
-    /// From environment
-    pub from_environment: MockEnvironmentName,
-    /// To environment
-    pub to_environment: MockEnvironmentName,
-    /// Promoted by user ID
-    pub promoted_by: String,
-    /// Approved by user ID (if applicable)
-    pub approved_by: Option<String>,
-    /// Status
-    pub status: PromotionStatus,
-    /// Timestamp
-    pub timestamp: chrono::DateTime<chrono::Utc>,
-    /// Comments
-    pub comments: Option<String>,
-    /// GitOps PR URL if created
-    pub pr_url: Option<String>,
-    /// Additional metadata (e.g., config changes diff)
-    #[serde(default)]
-    pub metadata: HashMap<String, serde_json::Value>,
-}
+// `PromotionHistoryEntry` is re-exported from `mockforge_foundation::workspace_promotion`.
+pub use mockforge_foundation::workspace_promotion::PromotionHistoryEntry;
 
 #[cfg(test)]
 mod tests {
